Update query-pubmed-grants.R
based on the meeting with Tom today, I made the changes to the query pubmed grants script in order to add abstracts as wikis on the publication entities

after checking how the first 5 went in testing, it became clear that running the script would not duplicate the entities, and i went forward with running it for all entities and adding the abstracts to the publications already in synapse

after triple checking i believe these were all of the final edits we made for it to run successfully, but let me know if anything is incorrect or missing. thank you SO much again for your help!

Signed-off-by: Melissa Klein <165932110+melissarklein@users.noreply.github.com>